fontconfig_reinit {gdtools}	R Documentation

reload Fontconfig configuration

Description

This function can be used to make fontconfig reload font configuration files.

Usage

fontconfig_reinit()

Note

Fontconfig is not used anymore and that function will be deprecated in the next release.

Author(s)

Paul Murrell

[Package gdtools version 0.2.2 Index]