glyphs_match {gdtools}	R Documentation

Validate glyph entries

Description

Determines if strings contain glyphs not part of a font.

Usage

glyphs_match(x, fontname = "sans", bold = FALSE, italic = FALSE,
  fontfile = "")

Arguments

x
Character vector of strings

fontname
Font name

bold, italic
Is text bold/italic?

fontfile
Font file

Value

a logical vector, if a character element is containing at least a glyph that can not be matched in the font table, FALSE is returned.

Examples

glyphs_match(letters)
glyphs_match("\u265E", bold = TRUE)

[Package gdtools version 0.2.2 Index]