match_family {gdtools}	R Documentation

Find best family match with systemfonts

Description

match_family() returns the best font family match.

Usage

match_family(font = "sans", bold = TRUE, italic = TRUE,
  debug = NULL)

Arguments

font
family or face to match.

bold
Wheter to match a font featuring a bold face.

italic
Wheter to match a font featuring an italic face.

debug
deprecated

Examples

match_family("sans")
match_family("serif")

[Package gdtools version 0.2.2 Index]